requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
grpcio==1.66.1
grpcio-reflection==1.66.1
grpcio-testing==1.66.1
grpcio-tools==1.66.1
gunicorn==21.2.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
packaging==24.1
pluggy==1.5.0
protobuf==5.28.2
pytest==8.3.3
pytest-dotenv==0.5.2
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.1.0
urllib3==2.2.3
Werkzeug==3.0.4